♻️ Refactor: 푸터에서에서 링크 사용 부분만 클라이언트 컴포넌트로 분리
