+++ src/app/(root)/layout.tsx
@@ -1,4 +1,4 @@
-import Footer from '@common/components/layout/Footer';
+import Footer from '@common/components/layout/Footer/Footer';
 
 const MainLayout = ({
   children,
@@ -6,7 +6,7 @@ const MainLayout = ({
   children: React.ReactNode;
 }>) => {
   return (
-    <div className="flex min-h-screen flex-col">
+    <div className="flex min-h-screen w-[1440px] flex-col">
       <main className="flex-1">{children}</main>
       <Footer />
     </div>
